Chapter 19 challenge 2.
This challenge consisted in keeping track of which strings are
allocated, and which are simple references from the source code. To
solve this, I simply added an 'owning' field. (there may be a few memory
leaks, which I don't feel looking for right now).
The flexible array member code is all commented out.